+++ R/raws_filterDate.R
@@ -101,9 +101,13 @@ raws_filterDate <- function(
   if (dateRange[1] > rawsObject$data$datetime[length(rawsObject$data$datetime)] |
       dateRange[2] < rawsObject$data$datetime[1]) {
     
-    warning(sprintf("rawsObject (%s) does not contain requested date range", rawsObject$meta$siteName))
+    message(sprintf(
+      "rawsObject (%s) does not contain requested date range", 
+      rawsObject$meta$siteName
+    ))
+    
     data <- rawsObject$data[0,]
-
+    
   } else {
     
     data <-